Move assert* from utils into assert_utils
For historical reasons, the `utils` module is a grab bag of various helper functions.
Now that we have a dedicated module `assert_utils` for assert helpers, move all these assertion helpers there.
